Fix_commit

+++ Vecteur_Surcharge_TP_CPP/Vecteur_Surcharge_TP_CPP/CVect2D.cpp
@@ -1,41 +1,65 @@
 #include "CVect2D.h"
 
+//BUT:	 Initialisé les valeurs du vecteur
+//ENTREE:2 floats
+//SORTIE:le vecteur initialisé
 CVect2D::CVect2D(float flt_x, float flt_y)
 {
 	this->flt_x = flt_x;
 	this->flt_y = flt_y;
 }
 
+//BUT:   Initialisé avec constructeur de copie
+//ENTREE:Un vecteur
+//SORTIE:Un vecteur
 CVect2D::CVect2D(const CVect2D& v)
 {
 	this->flt_x = v.flt_x;
 	this->flt_y = v.flt_y;
 }
 
+//BUT:   Libéré la mémoire
+//ENTREE:/
+//SORTIE:/
 CVect2D::~CVect2D()
 {
 }
 
-int CVect2D::getX() const
+//BUT:   Obtenir la valeur de X
+//ENTREE:/
+//SORTIE:X
+float CVect2D::getX() const
 {
 	return this->flt_x;
 }
 
-int CVect2D::getY() const
+//BUT:   Obtenir la valeur de Y
+//ENTREE:/
+//SORTIE:Y
+float CVect2D::getY() const
 {
 	return this->flt_y;
 }
 
+//BUT:   Set la valeur de X
+//ENTREE:Un float
+//SORTIE:Une nouvelle valeur pour X
 void CVect2D::setX(float x)
 {
 	this->flt_x = x;
 }
 
+//BUT:   Set la valeur de Y
+//ENTREE:Un float
+//SORTIE:Une nouvelle valeur pour Y
 void CVect2D::setY(float y)
 {
 	this->flt_y = y;
 }
 
+//BUT:   Additionner deux vecteurs
+//ENTREE:Un vecteur
+//SORTIE:Un vecteur
 CVect2D CVect2D::SomVect(const CVect2D& v) const
 {
 	CVect2D vTot;
@@ -44,6 +68,9 @@ CVect2D CVect2D::SomVect(const CVect2D& v) const
 	return vTot;
 }
 
+//BUT:   Soustraire deux vecteurs
+//ENTREE:Un vecteur
+//SORTIE:Un vecteur
 CVect2D CVect2D::SousVect(const CVect2D& v) const
 {
 	CVect2D vTot;
@@ -52,6 +79,9 @@ CVect2D CVect2D::SousVect(const CVect2D& v) const
 	return vTot;
 }
 
+//BUT:   Multiplier deux vecteurs
+//ENTREE:Un vecteur
+//SORTIE:Un float
 float CVect2D::MultVect(const CVect2D& v) const
 {
 	float flt_tot;
@@ -59,6 +89,9 @@ float CVect2D::MultVect(const CVect2D& v) const
 	return flt_tot;
 }
 
+//BUT:   Multiplier un vecteur avec un scalaire
+//ENTREE:Un float
+//SORTIE:Un vecteur
 CVect2D CVect2D::MultVectScal(float flt_scal) const
 {
 	CVect2D vTot;
@@ -67,6 +100,9 @@ CVect2D CVect2D::MultVectScal(float flt_scal) const
 	return vTot;
 }
 
+//BUT:   Vérifier si deux vecteurs sont égaux
+//ENTREE:Deux vecteurs
+//SORTIE:Un bool
 bool coincide2D(const CVect2D& v1, const CVect2D& v2)
 {
 	if (v1.flt_x == v2.flt_x && v1.flt_y == v2.flt_y) {
@@ -75,6 +111,9 @@ bool coincide2D(const CVect2D& v1, const CVect2D& v2)
 	return false;
 }
 
+//BUT:   Afficher les données du vecteur
+//ENTREE:/
+//SORTIE:Affiche les données
 void CVect2D::AfficheLog()const
 {
 	cout << "X : " << this->flt_x << endl;
